refactor(dstackup): split the 925-line main.rs into focused modules
main.rs was one 925-line file. Break it by concern:
- cli.rs     — clap definitions (Cli, Command, InstallOpts)
- install.rs — cmd_install + its helpers (key provider, image pin, port
               preflight, readiness probes)
- destroy.rs — cmd_destroy
- state.rs   — install-state persistence + the atomic write helper
- systemd.rs — unit management + the sanitized tool() spawner
- cid.rs     — CID-window allocation (+ its tests)
- main.rs    — now ~70 lines: module decls, main() dispatch, cmd_status

Also collapses the Command::Install 20-field destructure-then-rebuild into
`Install(InstallOpts)` via a clap `Args` derive (the duplication the first
review flagged). Pure reorganization, no behavior change: build, clippy
(-D warnings -D expect_used -D unwrap_used), fmt, the CID tests, and the full
`install --help` surface (all 20 flags + subcommands) are unchanged.
